+++ backend/src/main/resources/application.properties
@@ -1,1 +1,5 @@
-spring.application.name=warehouse
+spring.application.name=backend
+
+spring.datasource.url=${db-connection}
+spring.datasource.username=${db-user}
+spring.datasource.password=${db-pwd}
